requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
PyPDF2==3.0.1
langchain==0.0.340
numpy==1.24.3
qdrant-client==1.7.0
requests==2.31.0
groq==0.4.1
tavily-python==0.3.3
Werkzeug==2.3.7